chore(suite): moved global modals into OnboardingLayout
